==> 1.1 Table1.Rmd <==
---
title: "eTable1 (supplementary)"
author: "Yuqi Zhang"
date: "2025-08-15"
output:
  word_document: default
  html_document:
    keep_md: true
---

```{r, echo=FALSE}
load("/Users/yuqzha/LIBRO1_BCSTO/IC_timing_prognosis/data/cohorts.RData")
```

```{r, echo=FALSE, message=FALSE}
library(tidyverse)
library(knitr)
library(kableExtra)
```

## Create variable list & label mapping
```{r, echo=FALSE}
# Define your variable lists
factors1 <- c("year1","period", "parity","age_1st_child", "Education_2cat", "FH_BC_bl_50") 
tc <- c("T_pad_c", "N_pad_positive", "grade", "ER_negative", "PR_negative", "HER2_positive", "ki67", "histology")
trt <- c("post_chemo", "post_radio", "post_endo") 
factors2 <- c("pd_BIRADs","pd_3group","bmi_cat", "Alcohol_day", "smoking_packyears","risk_breastcancer_own_10yr","GRS_BC", "protein_LOF_rare_variants_exonic", "pinkcat","BOADICEA")

# Define ONLY continuous variables (everything else will be treated as factors automatically)
continuous_vars <- c("Alcohol_day", "smoking_packyears")

# Combine all variables you want to analyze
all_analysis_vars <- c(factors1, tc, trt, factors2)

# Create variable labels for better presentation
variable_labels <- c(
  # Basic demographics and history
  "year1" = "Year of diagnosis",
  "period" = "Diagnosis period",
  "parity" = "Parity", 
  "age_1st_child" = "Age at first child",
  "Education_2cat" = "Education level",
  "FH_BC_bl_50" = "Family history of breast cancer",
  
  # Tumor characteristics
  "T_pad_c" = "Tumor size (T stage)",
  "N_pad_positive" = "Lymph node status",
  "grade" = "Tumor grade",
  "ER_negative" = "ER negative",
  "PR_negative" = "PR negative", 
  "HER2_positive" = "HER2 positive",
  "ki67" = "Ki-67 index",
  "histology" = "Histological type",
  
  # Treatment variables
  "post_chemo" = "Post-operative chemotherapy",
  "post_radio" = "Post-operative radiotherapy", 
  "post_endo" = "Post-operative endocrine therapy",
  
  # Additional risk factors
  "pd_BIRADs" = "Breast density (BI-RADS)",
  "pd_3group" = "Breast density (3 groups)",
  "bmi_cat" = "BMI",
  "Alcohol_day" = "Alcohol consumption (g/day)",
  "smoking_packyears" = "Smoking (pack-years)",
  "risk_breastcancer_own_10yr" = "10-year breast cancer risk score",
  "GRS_BC" = "Genetic risk score",
  "protein_LOF_rare_variants_exonic" = "Protein-truncating variants",
  "pinkcat" = "PINK1 PTV",
  "BOADICEA" = "BOADICEA PTV"
)
```

```{r, echo=FALSE}
# source("FUN_desc_table.R") # does not work with merging rows, so changed to flextable

source("FUN_desc_flextable.R")
```

## eTable 1. Stockholm women cohort
```{r, echo=FALSE}
create_descriptive_flextable(
  data = cohort1,
  group_var = "detection_mode",
  variables = c(factors1, tc, trt),
  continuous_vars = NULL,
  use_median = TRUE,
  digits = 2,
  exact_tests = FALSE,
  variable_labels = variable_labels  # Add this parameter
)
```

## eTable 1. pKARMA and KARMA cohorts
```{r}
create_descriptive_flextable(
  data = cohort2,
  group_var = "detection_mode",
  variables = c(factors1, tc, trt, factors2),
  continuous_vars = continuous_vars,
  use_median = TRUE,
  digits = 2,
  exact_tests = FALSE,
  variable_labels = variable_labels  # Add this parameter
)
```
