==> 3.2 OR by density.Rmd <==
---
title: "3.2 OR by density"
author: "Yuqi Zhang"
date: "2025-04-24"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
```

## by density

```{r}
factors2 <- c("T_pad_c_2cat", "N_pad_positive", "grade", "ER_negative", "PR_negative", "HER2_positive", "ki67", "histology")

cohort <- cohort2 %>%
  filter(abs(bc_diagdate - pd_date) <= 364.25*2) |> # mammogram date within 2y
  # filter(!is.na(pd_3group)) |> 
  mutate(
    histology = case_when(
      histology %in% c("duc+lob","lobular") ~ "lobular",
      histology %in% c("other","other invasive") ~ "other",
      TRUE ~ histology),
    histology = factor(histology, levels = c("ductal", "lobular", "other")),
    across(all_of(factors2), factor)) # |> filter(M_stage != 1)
```

## strata by pd
```{r}
cohort <- cohort |> 
  filter(pd_3group %in% c("[0,20]", "(40,100]")) |> 
  mutate(pd_3group = factor(pd_3group, levels = c("[0,20]","(40,100]"))) 
```

```{r}
# Then run the analysis
source("FUN_uni_logistic.R")
results <- univariable_logistic(data = cohort, 
                                         group_var = "pd_BIRADs_3cat",#"pd_3group", #"sc_ic", 
                                         outcome_var =  "sc_ic", 
                                         predictor_vars = factors2) 
print(results)
```

## logistic, age-adjusted
```{r}
# Check what terms are actually generated
print(unique(results$term))

source("FUN_logistic_loop.R")
results <- logis_loop(data = cohort,
                                group_var = "pd_BIRADs_3cat",#"pd_3group", #"sc_ic",
                                outcome_var =  "sc_ic",
                                predictor_vars = factors2,
                                covariates = c("age_at_diagnosis")) 
print(results)
```
## filter results in terms
```{r}
source("OR_tc_filter_results.R")
```

### only low and high
```{r}
results_df <- filtered_results
source("OR_tc_compare_group.R")
```

### age plot
```{r}
OR_pd <- ggplot(filtered_results, 
                     aes(x = reorder(label, -order), y = est, color = association)) +
  theme_bw(base_size = 12) +
scale_y_log10(
    limits = c(0.25, 40),  # Adjust these limits based on your actual data range
    breaks = c(0.1, 0.2, 0.5, 1, 2, 5)) +
  geom_hline(yintercept = 1, linetype = 5, color = "grey") +
  geom_point(position = position_dodge(width = 0.5), shape = 17, size = 2) +
  geom_errorbar(aes(ymin = min95, ymax = max95), 
                position = position_dodge(width = 0.5),
                width = 0.2, size = 0.5) +
  # Add text with formatted OR and CI
  geom_text(
    aes(y = 5,
      label = sprintf("%.2f (%.2f, %.2f)", est, min95, max95),
      color = association),
    position = position_dodge(width = 0.5),
    hjust = 0,  # Align text to the left
    vjust = 0.5,  # Center vertically
    size = 3
  ) +
  labs(title = "IntCa vs. ScrCa phenotype by breast density (BIRADS)",
       x="",
  y="Odds ratio") +
  coord_flip() +
  facet_grid(.~ pd_BIRADs_3cat) + # pd_3group, labeller = labeller(pd_3group = pd_3group_label)) +
  scale_color_manual(values = c(
    "Positive" = "maroon",
    "Negative" = "forestgreen",
    "NS" = "grey40"
  ), guide = guide_legend(title = "", order = 1)) +
    theme(
    panel.grid.minor = element_blank(),
    panel.grid.major.x = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.spacing.y = unit(0.2, "cm"),
    legend.justification = "center",      # Align entire legend box to left
    legend.box.just = "center",           # Align legend box content to left
    plot.margin = margin(10, 10, 10, 10) # Add margins to ensure legend fits
  )

print(OR_pd)
```
```{r}
OR_pd_age <- OR_pd

OR_pd_compare <- OR_pd / OR_pd_age
```

## strata by calc
```{r}
# Then run the analysis
results2_strata <- univariable_logistic(data = cohort, 
                                         group_var = "calcs_bi", 
                                         outcome_var = "sc_ic", 
                                         predictor_vars = factors2) 
print(results2_strata)
# Check what terms are actually generated
print(unique(results2_strata$term))
```
```{r}
## run chunk line 77
calc_label <- c(
  "0" = "No microcalcification",
  ">=1" = "Microcalcification >=1"
)

filtered_results <- filtered_results |> 
  filter(calcs_bi %in% c("0", ">=1"))

OR_calc <- ggplot(filtered_results, 
                     aes(x = reorder(label, -order), y = est, color = association)) +
  theme_bw(base_size = 12) +
scale_y_log10(
    limits = c(0.28, 300),  # Adjust these limits based on your actual data range
    breaks = c(0.1, 0.2, 0.5, 1, 2, 5)) +
  geom_hline(yintercept = 1, linetype = 5, color = "grey") +
  geom_point(position = position_dodge(width = 0.5), shape = 17, size = 2) +
  geom_errorbar(aes(ymin = min95, ymax = max95), 
                position = position_dodge(width = 0.5),
                width = 0.2, size = 0.5) +
  # Add text with formatted OR and CI
  geom_text(
    aes(y = 13,
      label = sprintf("%.2f (%.2f, %.2f)", est, min95, max95),
      color = association),
    position = position_dodge(width = 0.5),
    hjust = 0,  # Align text to the left
    vjust = 0.5,  # Center vertically
    size = 3.5
  ) +
  labs(title = "OR of interval cancer vs. screen-detected cancer",
       x="",
  y="Odds ratio") +
  coord_flip() +
  facet_grid(.~calcs_bi, labeller = labeller(calcs_bi = calc_label)) +
  scale_color_manual(values = c(
    "Positive" = "maroon",
    "Negative" = "forestgreen",
    "NS" = "grey40"
  ), guide = guide_legend(title = "", order = 1)) +
    theme(
    panel.grid.minor = element_blank(),
    panel.grid.major.x = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.spacing.y = unit(0.2, "cm"),
    legend.justification = "center",      # Align entire legend box to left
    legend.box.just = "center",           # Align legend box content to left
    plot.margin = margin(10, 10, 10, 10) # Add margins to ensure legend fits
  )

print(OR_calc)
```
```{r}
OR_tc_by_calcs <- filtered_results
```

## strata by masses
```{r}
# Then run the analysis

cohort <- cohort |> 
  mutate(masses_mean2cat = case_when(
      masses == 0 ~ 0, 
      masses >= 1 ~ 1),
    masses_mean2cat = factor(masses_mean2cat, labels = c("0", ">=1")))

results2_strata <- univariable_logistic(data = cohort, 
                                         group_var = "sc_ic",  
                                         outcome_var = "masses_mean2cat",
                                         predictor_vars = factors2) 
print(results2_strata)
# Check what terms are actually generated
print(unique(results2_strata$term))
```
```{r}
# run line 77 

mass_label <- c(
  "0" = "No mass",
  ">=1" = "Mass >=1"
)

filtered_results <- filtered_results |> 
  filter(!is.na(sc_ic))

OR_mass <- ggplot(filtered_results, 
                     aes(x = reorder(label, -order), y = est, color = association)) +
  theme_bw(base_size = 12) +
scale_y_log10(
    limits = c(0.14, 300),  # Adjust these limits based on your actual data range
    breaks = c(0.1, 0.2, 0.5, 1, 2, 5)) +
  geom_hline(yintercept = 1, linetype = 5, color = "grey") +
  geom_point(position = position_dodge(width = 0.5), shape = 17, size = 2) +
  geom_errorbar(aes(ymin = min95, ymax = max95), 
                position = position_dodge(width = 0.5),
                width = 0.2, size = 0.5) +
  # Add text with formatted OR and CI
  geom_text(
    aes(y = 13,
      label = sprintf("%.2f (%.2f, %.2f)", est, min95, max95),
      color = association),
    position = position_dodge(width = 0.5),
    hjust = 0,  # Align text to the left
    vjust = 0.5,  # Center vertically
    size = rel(1.5)
  ) +
  labs(title = "OR of masses (mean of both sides) >=1 vs. 0",
       x="",
  y="Odds ratio") +
  coord_flip() +
  facet_grid(.~sc_ic, labeller = labeller(sc_ic = scic_label)) + #masses_bi, labeller = labeller(masses_bi = mass_label)
  scale_color_manual(values = c(
    "Positive" = "maroon",
    "Negative" = "forestgreen",
    "NS" = "grey40"
  ), guide = guide_legend(title = "", order = 1)) +
    theme(
    panel.grid.minor = element_blank(),
    panel.grid.major.x = element_blank(),
    legend.position = "bottom",
    legend.box = "vertical",
    legend.spacing.y = unit(0.2, "cm"),
    legend.justification = "center",      # Align entire legend box to left
    legend.box.just = "center",           # Align legend box content to left
    plot.margin = margin(10, 10, 10, 10) # Add margins to ensure legend fits
  )

print(OR_mass)
```
```{r}
OR_tc_by_mass <- filtered_results
```
