==> 4 mortality by year.Rmd <==
---
title: "4 mortality by year"
author: "Yuqi Zhang"
date: "2025-08-26"
output:
  word_document:
    keep_md: true
  pdf_document: default
  html_document: default
always_allow_html: true
---
To estimate the HR of BC death associated with factors (some from cohort1, and mammogram-related factors from cohort 2)
Strata: IntCa or ScrCa
factors: factors1 from cohort1, factors2 from cohort
adjustment: age, additionally TC

steps:
1.set up the data (define outcome and time: death_BC, time_var, factors)
2.model cohort 1 and 2 separately

note: 
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, include=FALSE, message=FALSE}
library(tidyverse)
library(survival)
library(survminer)
library(ggplot2)
library(dplyr)
library(purrr)
library(splines)
library(haven) # to handle haven_labelled variables
library(flextable)
library(officer)
```

```{r}
load("/Users/yuqzha/LIBRO1_BCSTO/IC_timing_prognosis/data/cohorts.RData")
```


## define outcome: BC-specific death
```{r}
cohort1 <- cohort1 |> 
  mutate(
         time_start = diadat, # enter at diagnosis
         time_end = ymd("2023-12-31"),
         time_to_death = case_when(
             !is.na(death_date) & death_date >= time_start ~ as.numeric(death_date - time_start),
             TRUE ~ NA_real_),
         time_to_emig = case_when(
             !is.na(emigration_last_date) & emigration_last_date > time_start ~ as.numeric(emigration_last_date - time_start),
             TRUE ~ NA_real_),
         death_BC =  case_when(
           !is.na(death_date)  & substr(death_cause, 1,3) %in% c("C50") ~ 1,
           # !is.na(death_date)  & substr(death_cause, 1,3) %in% c("C50", "174") & time_to_death <= 365*10 ~ 1, 
            # 10-year mortality
             TRUE ~ 0),
      # death_BC = ifelse(is.na(death_cause), 0, # for competing risk
         #                 ifelse(substr(death_cause, 1,3) %in% c("C50", "170"), 1, 2)),
         time_var = pmin(
           ifelse(is.na(time_to_death), Inf, time_to_death),
           ifelse(is.na(time_to_emig), Inf, time_to_emig),
           as.numeric(time_end - time_start),
           na.rm = TRUE)) #365*10
```

```{r}
# as of Aug 3 2025
factors1 <- c("year1","period", "parity","age_1st_child", "Education_2cat", "Breast_n_bi","FH_BC_bl_50","FH_BC_bl_40") 
tc <- c("T_pad_c_2cat",  "N_pad_positive", "ER_negative", "grade")
trt <- c("post_chemo") #, "post_radio","post_endo") # post_endo


#factors1 <- c("year1")
covariate1 <- "age_at_diagnosis"
covariate2 <- c("age_at_diagnosis", tc)
covariate3 <- c("age_at_diagnosis", tc, trt) 

covariate1 <- "age_at_diagnosis"
covariate2 <- c("age_at_diagnosis", "insitu","T_pad_c_2cat",  "N_pad_positive", "ER_negative","grade")
covariate3 <- c("age_at_diagnosis", "T_pad_c_2cat",  "N_pad_positive", "ER_negative", "grade","post_chemo") 
# load function of coxph loop
source("FUN_coxph_loop.R")
```

# factors based on cohort 1
```{r}
# Convert columns to factors before analysis
cohort <- cohort1 %>%
  #filter(M_stage == 0 | is.na(M_stage)) |> #
  #filter(insitu != 1) |> 
  mutate(across(all_of(c(factors1, tc, trt)), factor)) |> 
  filter(detection_mode == "IntCa")

# Then run the analysis
cox0 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors1,
                   covariates = NULL,
                   group_var = "year1") |> 
  mutate(cohort = "Cohort 1",
         model = "Model 0")

# age adjusted
cox1 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors1,
                   covariates = covariate1,
                   group_var = "year1") |> 
    mutate(cohort = "Cohort 1",
         model = "Model 1")

# age and TC adjusted
cox2 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors1,
                   covariates = covariate2,
                   group_var = "year1") |> 
    mutate(cohort = "Cohort 1",
         model = "Model 2")

# age, TC and trt adjusted
cox3 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors1,
                   covariates = covariate3,
                   group_var = "year1") |> 
    mutate(cohort = "Cohort 1",
         model = "Model 3")

result1 <- bind_rows(cox1, cox2, cox3)

```

## cohort 2
```{r}

## as of Aug 3
factors2 <- c("pd_3group","pd_BIRADs","pd_BIRADs_3cat","pd_BIRADs_2cat","bmi_cat", "Alcohol_day", "smoking_packyears", "risk_breastcancer_own_10yr","GRS_BC", "protein_LOF_rare_variants_exonic", "pinkcat","BOADICEA")

#factors2 <- c("bmi_cat", "Alcohol_day", "smoking_packyears", "risk_breastcancer_own_10yr","GRS_BC", "protein_LOF_rare_variants_exonic", "pinkcat","BOADICEA")

# Convert columns to factors before analysis
cohort2 <- cohort2 |> 
  #filter(pd <= 40) |> 
  #filter(M_stage == 0 | is.na(M_stage)) |> #
  #filter(insitu != 1) |> 
  mutate(
    #time_start = pmax(bc_diagdate, studyentry_date), #delayed entry to start from study entry date or diagnosis date, whichever came later
    time_start = bc_diagdate, #no delay entry, enter at diagnosis
    time_end = ymd("2022-05-07"),
    time_to_death = case_when(
             !is.na(death_date) & death_date >= time_start ~ as.numeric(death_date - time_start),
             TRUE ~ NA_real_),
         # time_to_emig = case_when(
         #     !is.na(emigration_last_date) & emigration_last_date > time_start ~ as.numeric(emigration_last_date - time_start),
         #     TRUE ~ NA_real_),
         death_BC =  case_when(
           !is.na(death_date)  & substr(death_cause, 1,3) %in% c("C50") ~ 1,
           # !is.na(death_date)  & substr(death_cause, 1,3) %in% c("C50", "174") & time_to_death <= 365*10 ~ 1, 
            # 10-year mortality
             TRUE ~ 0),
         time_var = pmin(
           ifelse(is.na(time_to_death), Inf, time_to_death),
           as.numeric(time_end - time_start),
           na.rm = TRUE)) #365*10

# re-order some reference groups   
cohort <- cohort2 |> 
    filter(detection_mode == "IntCa") |> 
    mutate(
    pd_BIRADs = factor(pd_BIRADs,
                       levels = c("(18,49]", "[0,2]", "(2,18]","(49,100]")),
    pd_BIRADs_3cat = factor(pd_BIRADs_3cat, 
                            levels =  c("C:(18,49]", "A+B:[0,18]", "D:(49,100]")),
    pd_BIRADs_2cat = case_when(
      pd_BIRADs_3cat == "A+B:[0,18]" ~ "A+B:[0,18]",
      pd_BIRADs_3cat %in% c("C:(18,49]", "D:(49,100]") ~ "C+D:(18,100]",
      TRUE ~ NA
    ),
    pd_3group = factor(pd_3group,
                       levels = c("(20,40]","[0,20]", "(40,100]")),
    across(c(protein_LOF_rare_variants_exonic, pinkcat, BOADICEA), factor)) 


# Then run the analysis
cox0 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors2,
                   covariates = NULL,
                   group_var = "year1") |> 
  mutate(cohort = "Cohort 2",
         model = "Model 0")

# age adjusted
cox1 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors2,
                   covariates = covariate1,
                   group_var = "year1") |> 
  mutate(cohort = "Cohort 2",
         model = "Model 1")

# age and TC adjusted
cox2 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors2,
                   covariates = covariate2,
                   group_var = "year1") |> 
  mutate(cohort = "Cohort 2",
         model = "Model 2")

# age, TC and trt adjusted
cox3 <- coxph_loop(data = cohort, outcome_var = "death_BC", time_var = "time_var",
                   predictor_vars = factors2,
                   covariates = covariate3,
                   group_var = "year1") |> 
  mutate(cohort = "Cohort 2",
         model = "Model 3")

result2 <- bind_rows(cox1, cox2, cox3)
```

## result bind, order, and pivot
```{r}
# create order of factors, choose one set of model from each cohort and bind
order <- bind_rows(
  result1 %>% filter(model == "Model 1" & year1 == "Year 1") %>% select(predictor_var, Level),
  result2 %>% filter(model == "Model 1" & year1 == "Year 1") %>% select(predictor_var, Level)
)  |> 
  mutate(order = row_number())

# join results, add order
result <- bind_rows(result1, result2) |> 
  left_join(order, by = c("predictor_var", "Level"))
```

```{r}
# First, reshape your data to wide format
result <- result |> 
  mutate(group = paste(year1, model, sep = "_"))

dat <- result |> 
  select(predictor_var, Level,order, group, est)

dat_wide <- dat %>%
  pivot_wider(names_from = group, 
              values_from = est,
              id_cols = c(predictor_var, Level, order))

dat_wide <- dat_wide |> 
  select(predictor_var, Level, order, starts_with(c("Year 1_", "Year 2_")))


```

```{r}
# Define variable labels
variable_labels <- c(
  "year1" = "Year of diagnosis",
  "period" = "Diagnosis period",
  "parity" = "Parity", 
  "age_1st_child" = "Age at first child",
  "Education_2cat" = "Education level",
  "FH_BC_bl_50" = "Family history of breast cancer",
  "pd_BIRADs" = "Breast density (BI-RADS)",
  "pd_3group" = "Breast density (3 groups)",
  "pd_BIRADs_3cat" = "Breast density (3 categories)",
  "bmi_cat" = "BMI",
  "Alcohol_day" = "Alcohol consumption (g/day)",
  "smoking_packyears" = "Smoking (pack-years)",
  "risk_breastcancer_own_10yr" = "10-year breast cancer risk score"
)

# Add variable labels to dat_wide
dat_wide <- dat_wide %>%
  mutate(
    factor_labeled = ifelse(predictor_var %in% names(variable_labels),
                           variable_labels[predictor_var],
                           predictor_var)
  ) %>%
  select(factor_labeled, Level, everything(), -predictor_var, -order)

# Create flextable with your original code (modified)
ft <- dat_wide %>%
  flextable::flextable() %>%
  # Add the main header row for ScrCa and IntCa
  flextable::add_header_row(values = c("", "", rep("Year 2", 3), rep("Year 1", 3)), top = TRUE) %>%
  # Merge cells for same predictor variables
  flextable::merge_v(j = "factor_labeled", part = "body") %>%
  # Merge cells for headers
  flextable::merge_h(part = "header", i = 1) %>%
  flextable::merge_h(part = "header", i = 2) %>%
  # Styling
  flextable::theme_booktabs() %>%
  flextable::bold(part = "header") %>%
  flextable::align(align = "center", part = "header") %>%
  flextable::align(align = "left", j = c("factor_labeled", "Level")) %>%
  # Center-align the merged predictor variable column vertically
  flextable::valign(j = "factor_labeled", valign = "center", part = "body") %>%
  flextable::autofit()

# Print the table
ft
```
